First working version
Has all the wanted features, but there are a few cleanup things.
Add users file to config, so the two parts of the app can live in
separate places.
Make the slack_api part that runs on flask work with a flask production
environment.
Add more info to the README