Fix for statuses using Unicode characters
Some Steam game names have Unicode characters in them, which was causing
the app to crash when trying to post that to Slack. The issue was with
passing the Unicode string to the requests library. I still don't
understand why it didn't work (isn't Python 3 supposed to "just work"
with Unicode?), but this Stack Overflow answer pointed me in the right
direction: https://stackoverflow.com/q/9942594.