Update Slack Oauth to their v2
This isn't Oauth V2, since they were already using that, this is their
version of Oauth V2, V2. Not many changes, but good to be updated, and
I added a change so that the app doesn't need to be configured in Slack
with a redirect URI. Now the app passes that in everywhere.