Rework status daemon section to integrate better
Update the status daemon to use all the now async methods of the
library. Convert all the daemon methods to also be async too.
Break up the main method so that the parts that the API server handles
are in the main(), and everything else is not.
Add a startup of the daemon to the API server so that multiple processes
to handle this aren't needed. The status daemon can still be run solo,
but it isn't needed now.