Add defaults for gunicorn worker count and binding
More workers for an API that is almost never hit is pointless. And the
way I have it, each worker runs its own status daemon, which isn't very
useful.

Change the default bind address to be IPv6, which on most Linux kernels
is set to bind to both IPv4 and 6.